blockchain/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.2.1
itsdangerous==2.2.0
MarkupSafe==3.0.2
psycopg2-binary==2.9.11
gunicorn==25.3.0
python-dotenv==1.2.2
flask-cors==6.0.1
